feat: add support for 64-bit symbol lookup tables (#22)
To overcome the 4 GiB file size limit some operating system like Solaris
11.2 and GNU use a variant lookup table. Instead of 32-bit integers,
64-bit integers are used in the symbol lookup tables. The string
"/SYM64/" instead "/" is used as identifier for this table[0].

A 32-bit archive symbol table has a zero length name, so ar_name
contains the string "/" padded with 15 blank characters on the right. A
64-bit archive symbol table sets ar_name to the string "/SYM64/", padded
with 9 blank characters to the right.

[0]: https://docs.oracle.com/cd/E36784_01/html/E36873/ar.h-3head.html